refactor!: switch to raw-first API, no implicit f32 conversion
PhaseReadings and all per-phase read methods now return raw register
values (u16, i32, i16) instead of f32.  This avoids lossy conversion,
eliminates mandatory FPU dependency, and gives callers full control
over when and how to convert to engineering units.

Conversion helpers in proto module are unchanged and remain available
for on-demand use:
  proto::voltage_raw_to_volts(raw)
  proto::current_raw_to_amps(raw)
  proto::power_raw_to_watts(hi, lo)
  proto::power_factor_raw_to_unitless(raw)
  proto::frequency_raw_to_hz(raw)
  proto::phase_angle_raw_to_degrees(raw)
  proto::chip_temperature_raw(raw)

BREAKING CHANGE: PhaseReadings field types changed from f32 to
u16/i32/i16. Per-phase read methods return raw integer types.

+++ src/readings.rs
@@ -6,36 +6,38 @@
 /// A full snapshot of the 3-phase measurements from a single
 /// [`Atm90e32::read_all_phases`] call.
 ///
-/// All values are in engineering units:
+/// All values are **raw register values** — lossless, no floating-point
+/// conversion. Use the [`proto`](crate::proto) helpers to convert to
+/// engineering units when needed:
 ///
-/// * `voltage` — RMS phase voltage, volts
-/// * `current` — RMS phase current, amps
-/// * `power` — active power, watts
-/// * `reactive` — reactive power, vars
-/// * `pf` — power factor, dimensionless in -1.0..=1.0
-/// * `frequency` — mains frequency, hertz
+/// | Field | Raw unit | Converter |
+/// |-------|----------|-----------|
+/// | `voltage` | hundredths of a volt (u16) | [`voltage_raw_to_volts`](crate::proto::voltage_raw_to_volts) |
+/// | `current` | thousandths of an amp (u16) | [`current_raw_to_amps`](crate::proto::current_raw_to_amps) |
+/// | `power` | signed 32-bit combined word | [`power_raw_to_watts`](crate::proto::power_raw_to_watts) |
+/// | `reactive` | signed 32-bit combined word | [`power_raw_to_watts`](crate::proto::power_raw_to_watts) |
+/// | `pf` | signed thousandths (i16) | [`power_factor_raw_to_unitless`](crate::proto::power_factor_raw_to_unitless) |
+/// | `frequency` | hundredths of a hertz (u16) | [`frequency_raw_to_hz`](crate::proto::frequency_raw_to_hz) |
+/// | `phase_angle` | tenths of a degree (u16) | [`phase_angle_raw_to_degrees`](crate::proto::phase_angle_raw_to_degrees) |
 ///
 /// Each three-element array is indexed in phase order A, B, C.
 ///
-/// The field layout is deliberately stable across patch releases so
-/// consumers can destructure by name safely.
-///
 /// [`Atm90e32::read_all_phases`]: crate::Atm90e32::read_all_phases
-#[derive(Debug, Clone, Copy, Default, PartialEq)]
+#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 pub struct PhaseReadings {
-    /// RMS phase voltage in volts, per phase (A, B, C).
-    pub voltage: [f32; 3],
-    /// RMS phase current in amps, per phase (A, B, C).
-    pub current: [f32; 3],
-    /// Active power in watts, per phase (A, B, C).
-    pub power: [f32; 3],
-    /// Reactive power in vars, per phase (A, B, C).
-    pub reactive: [f32; 3],
-    /// Power factor, dimensionless, per phase (A, B, C).
-    pub pf: [f32; 3],
-    /// Mains frequency in hertz.
-    pub frequency: f32,
-    /// Mean phase angle in degrees (0..360), per phase (A, B, C).
-    pub phase_angle: [f32; 3],
+    /// RMS phase voltage, per phase (A, B, C). Raw: hundredths of a volt.
+    pub voltage: [u16; 3],
+    /// RMS phase current, per phase (A, B, C). Raw: thousandths of an amp.
+    pub current: [u16; 3],
+    /// Active power, per phase (A, B, C). Raw: signed 32-bit combined hi+lo register words.
+    pub power: [i32; 3],
+    /// Reactive power, per phase (A, B, C). Raw: signed 32-bit combined hi+lo register words.
+    pub reactive: [i32; 3],
+    /// Power factor, per phase (A, B, C). Raw: signed thousandths (-1000..=1000).
+    pub pf: [i16; 3],
+    /// Mains frequency. Raw: hundredths of a hertz.
+    pub frequency: u16,
+    /// Mean phase angle, per phase (A, B, C). Raw: tenths of a degree.
+    pub phase_angle: [u16; 3],
 }
